AddPaymentMethod Package Interface
* AddPaymentMethod -> Topup, FinanceHome에서 사용하기 떄문에 의존성을 끊어주는게 의미가 있다
