+++ Finance/Package.swift
@@ -7,10 +7,13 @@ let package = Package(
     name: "Finance",
     platforms: [.iOS(.v14)],
     products: [
-        // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
             name: "AddPaymentMethod",
             targets: ["AddPaymentMethod"]
+        ),
+        .library(
+            name: "AddPaymentMethodImp",
+            targets: ["AddPaymentMethodImp"]
         ),
         .library(
             name: "FinanceEntity",
@@ -36,16 +39,21 @@ let package = Package(
     dependencies: [
         .package(name: "ModernRIBs", url: "https://github.com/DevYeom/ModernRIBs", .exact("1.0.1")),
         .package(path: "../Platform")
-        // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
     ],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "AddPaymentMethod",
             dependencies: [
                 "ModernRIBs",
+                "FinanceEntity",
+                .product(name: "RIBsUtil", package: "Platform")
+            ]
+        ),
+        .target(
+            name: "AddPaymentMethodImp",
+            dependencies: [
+                "ModernRIBs",
+                "AddPaymentMethod",
                 "FinanceEntity",
                 "FinanceRepository",
                 .product(name: "RIBsUtil", package: "Platform"),
