+++ Finance/Sources/AddPaymentMethodImp/AddPaymentMethodRouter.swift
@@ -6,6 +6,7 @@
 //
 
 import ModernRIBs
+import AddPaymentMethod
 
 protocol AddPaymentMethodInteractable: Interactable {
     var router: AddPaymentMethodRouting? { get set }
