+++ Finance/Sources/FinanceHome/FinanceHomeBuilder.swift
@@ -17,9 +17,11 @@ public protocol FinanceHomeDependency: Dependency {
     var cardOnFileRepository: CardOnFileRepository { get }
     
     var topupBuildable: TopupBuildable { get }
+    
+    var addPaymentMethodBuildable: AddPaymentMethodBuildable { get }
 }
 
-final class FinanceHomeComponent: Component<FinanceHomeDependency>, SuperPayDashboardDependency, CardOnFileDashboardDependency, AddPaymentMethodDependency {
+final class FinanceHomeComponent: Component<FinanceHomeDependency>, SuperPayDashboardDependency, CardOnFileDashboardDependency {
     
     var balance: ReadOnlyCurrentValuePublisher<Double> { superPayRepository.balance }
     var superPayRepository: SuperPayRepository { dependency.superPayRepository }
@@ -29,6 +31,8 @@ final class FinanceHomeComponent: Component<FinanceHomeDependency>, SuperPayDash
     var topupBaseViewController: ViewControllable
     
     var topupBuildable: TopupBuildable { dependency.topupBuildable }
+    
+    var addPaymentMethodBuildable: AddPaymentMethodBuildable { dependency.addPaymentMethodBuildable }
     
     init(
         dependency: FinanceHomeDependency,
@@ -62,14 +66,13 @@ public final class FinanceHomeBuilder: Builder<FinanceHomeDependency>, FinanceHo
         
         let superPayDashboard = SuperPayDashboardBuilder(dependency: component)
         let cardOnFileDasahboard = CardOnFileDashboardBuilder(dependency: component)
-        let addPaymentMethod = AddPaymentMethodBuilder(dependency: component)
         
         return FinanceHomeRouter(
             interactor: interactor,
             viewController: viewController,
             superPayDashboard: superPayDashboard,
             cardOnFileDashboard: cardOnFileDasahboard,
-            addPaymentMethod: addPaymentMethod,
+            addPaymentMethod: component.addPaymentMethodBuildable,
             topup: component.topupBuildable
         )
     }
