+++ Finance/Sources/TopupImp/TopupBuilder.swift
@@ -17,9 +17,10 @@ public protocol TopupDependency: Dependency {
     var topupBaseViewController: ViewControllable { get }
     var cardOnFileRepository: CardOnFileRepository { get }
     var superPayRepository: SuperPayRepository { get }
+    var addPaymentMethodBuildable: AddPaymentMethodBuildable { get }
 }
 
-final class TopupComponent: Component<TopupDependency>, TopupInteractorDependency, AddPaymentMethodDependency, EnterAmountDependency, CardOnFileDependency {
+final class TopupComponent: Component<TopupDependency>, TopupInteractorDependency, EnterAmountDependency, CardOnFileDependency {
     fileprivate var topupBaseViewController: ViewControllable {
         return dependency.topupBaseViewController
     }
@@ -29,6 +30,8 @@ final class TopupComponent: Component<TopupDependency>, TopupInteractorDependenc
     
     var selectedPayment: ReadOnlyCurrentValuePublisher<PaymentModel> { selectedPaymentStream }
     let selectedPaymentStream: CurrentValuePublisher<PaymentModel>
+    
+    var addPaymentMethodBuildable: AddPaymentMethodBuildable { dependency.addPaymentMethodBuildable }
     
     init(
         dependency: TopupDependency,
@@ -56,14 +59,13 @@ public final class TopupBuilder: Builder<TopupDependency>, TopupBuildable {
         let interactor = TopupInteractor(dependency: component)
         interactor.listener = listener
         
-        let addPaymentMethod = AddPaymentMethodBuilder(dependency: component)
         let enterAmount = EnterAmountBuilder(dependency: component)
         let cardOnFile = CardOnFileBuilder(dependency: component)
         
         return TopupRouter(
             interactor: interactor,
             viewController: component.topupBaseViewController,
-            addPaymentMethod: addPaymentMethod,
+            addPaymentMethod: component.addPaymentMethodBuildable,
             enterAmount: enterAmount,
             cardOnFile: cardOnFile
         )
